requirements.txt
# requirements.txt
# PyTorch avec CUDA pour GPU (ou version CPU)
# Pour CPU uniquement (plus léger):
torch==2.0.1+cpu
torchvision==0.15.2+cpu
-f https://download.pytorch.org/whl/torch_stable.html

# Pour GPU:
# torch==2.0.1+cu118
# torchvision==0.15.2+cu118

# Autres dépendances avec versions fixes
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
Pillow==9.5.0
PyYAML==6.0
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0

# Ajouter pour réduire la taille
pip==23.1.2
setuptools==67.8.0
wheel==0.40.0